requirements.txt
pytest
pytest-playwright
playwright
allure-pytest
setuptools
wheel
pytest-xdist